stable-fast-3d/requirements.txt
einops==0.7.0
jaxtyping==0.2.31
omegaconf==2.3.0
transformers==4.42.3
open_clip_torch==2.24.0
trimesh==4.4.1
numpy==1.26.4
huggingface-hub==0.23.4
rembg==2.0.57
pynanoinstantmeshes==0.0.3
fastapi
uvicorn[standard]
python-multipart
python-dotenv
